Analyzing Acquisition of Conrail Railways
Questions & Hints – Case Study #7
Conrail A&B, HBS 298-006, 298-095
Corporate Finance
Professor Opp
The Conrail A&B cases vividly illustrate the strategies and tactics employed by firms in financial
markets around M&A events. CSX and Conrail, the first and third largest railroads in the Eastern
United States, announce their intent to merge in a friendly two-tiered transaction. One week after
they announce the deal, Norfolk Southern, the second largest railroad in the East, makes an
unsolicited competing bid for Conrail. Over the next few months, both bidders get into a bidding
war, prompting market analysts to call this one of the nastiest takeover battles of the 1990s. You
will evaluate the bids made by CSX and Norfolk Southern, and make a recommendation to the
Conrail Board and shareholders.
In your analysis, you should address the following points. As always, your report should be written
as a recommendation to the managers/investors that touches upon all of the below key issues.
Conrail – A
1. Why does CSX want to buy Conrail?
2. Analyze the offer by CSX. How much should CSX be willing to pay, and is there any
evidence to suggest CSX might be overpaying? To answer these questions, you should
assume that “Gains in Operating Income” (i.e., synergies) are net of costs (e.g., COGS
and Capital Expenditures).
3. Analyze the structure of CSX’s offer for Conrail. Why did CSX make a two-tiered offer?
What effect does this structure have on the transaction? Why did CSX want to do the
first-tier transaction in two steps?
4. Analyze the other conditions and provisions of the deal, such as the “break-up” fee, the
“lock-up” option, and the “no talk” clause. What are the implications of these
provisions?
Conrail – B
5. Why did Norfolk Southern make a hostile bid for Conrail?
6. In a bidding war, who should be willing to pay more, Norfolk Southern or CSX? In
estimating how much each firm should be willing to pay, be sure to account for what
happens if the other firm wins the contest. [Note: Some of the numbers in Case B are
different than those presented in Case A. Use the numbers in Case A to answer the first
four questions above, and the numbers in Case B to answer this question.]
7. As a shareholder, would you vote to opt-out of the Pennsylvania anti-takeover statute?
Explain your reasoning. [Note: Please make sure to think about when you will actually
get paid as a shareholder under each of the different proposals.]
8. Provide sensitivity analyses for your estimates above.